modularize api and fix tests
